Use off_t type from sys/types for next_num

+++ next_i.c
@@ -1,7 +1,8 @@
+#include <sys/types.h>
 #include "next_i.h"
 
-int next_num(int i){
-	if((i%2) == 0){
+off_t next_num (off_t i) {
+	if ((i%2) == 0) {
 		return i/2;
 	}
 	return (3*i) + 1;
